Adds a 'basic-list-view.component' to detach from operator
