+++ client/src/app/shared/components/basic-list-view-table/basic-list-view-table.component.html
@@ -1,0 +1,49 @@
+<mat-drawer-container class="list-view-frame" *ngIf="columns && columnSet">
+    <div class="list-view-table-wrapper" (keydown)="onKeyDown($event)" (keyup)="onKeyUp($event)">
+        <os-sort-filter-bar
+            class="sort-filter-bar"
+            *ngIf="showFilterBar"
+            [totalCount]="totalCount"
+            [filterCount]="countFilter"
+            [filterService]="filterService"
+            [sortService]="sortService"
+            [searchFieldInput]="inputValue"
+            (searchFieldChange)="searchFilter($event)"
+        >
+            <div class="extra-info-slot">
+                <ng-content select=".filter-bar-info-slot"></ng-content>
+            </div>
+        </os-sort-filter-bar>
+
+        <pbl-ngrid
+            class="vscroll-list-view"
+            [attr.vScrollFixed]="vScrollFixed !== -1 ? vScrollFixed : false"
+            [attr.vScrollAuto]="vScrollFixed === -1"
+            [showHeader]="!showFilterBar || !fullScreen"
+            [showHeader]="false"
+            matCheckboxSelection="selection"
+            [dataSource]="dataSource"
+            [columns]="columnSet"
+            [hideColumns]="hiddenColumns"
+            (rowClick)="onSelectRow($event)"
+            rowClassUpdateFreq="ngDoCheck"
+        >
+            <!-- "row" has the view model -->
+            <!-- "value" has the property, that was defined in the columnDefinition -->
+            <!-- "col" has a column reference -->
+            <ng-container *ngTemplateOutlet="startColumnView"></ng-container>
+
+            <!-- No Results -->
+            <div *pblNgridNoDataRef class="pbl-ngrid-no-data">
+                <span>{{ 'No data' | translate }}</span>
+            </div>
+
+            <!-- Slot transclusion for the individual cells -->
+            <div #contentWrapper>
+                <ng-content class="ngrid-lg" select=".cell-slot"></ng-content>
+            </div>
+
+            <ng-container *ngTemplateOutlet="endColumnView"></ng-container>
+        </pbl-ngrid>
+    </div>
+</mat-drawer-container>
