+++ client/src/app/shared/components/basic-list-view-table/basic-list-view-table.component.scss
